revert to 3d engine only
