--- game.h
@@ -1,51 +1,0 @@
-#pragma once
-#include "stb_image.h"
-#include <GL/glew.h>
-#include <GLFW/glfw3.h>
-#include <iostream>
-#include <chrono>
-#include <algorithm>
-#include <list>
-#include <stdio.h>
-#include <assert.h>
-#include <string.h>
-#include <math.h>
-#include <vector>
-#include <array>
-#include <glm/glm.hpp>
-#include <glm/gtc/matrix_transform.hpp>
-
-#include "util/vec3d.h"
-#include "util/mat4.h"
-#include "util/camera.h"
-#include "util/blockAtlas.h"
-#include "util/triangle.h"
-#include "util/Process3D.h"
-#include "util/block.h"
-#include "util/GLFWRender.h"
-
-
-
-// include imgui
-#include "lib/imgui/imgui.h"
-#include "lib/imgui/backends/imgui_impl_glfw.h"
-#include "lib/imgui/backends/imgui_impl_opengl3.h"
-
-
-
-extern const char *stbi_failure_reason(void);
-extern void stbi_image_free(void *ptr);
-extern stbi_uc * stbi_load(const char *filename, int *x, int *y, int *comp, int req_comp);
-extern void stbi_set_flip_vertically_on_load(int flag);
-extern unsigned char *stbi_load(const char *filename, int *x, int *y, int *channels, int desired_channels);
-extern void stbi_image_free(void *ptr);
-
-
-
-void checkGLError(const char* location) {
-    GLenum error = glGetError();
-    if (error != GL_NO_ERROR) {
-        std::cerr << "OpenGL error at " << location << ": " << error << std::endl;
-    }
-}
-
